Fix: fix request body create account DTO

+++ src/main/java/com/example/coffee_shop_chain_management/service/AccountService.java
@@ -1,16 +1,31 @@
 package com.example.coffee_shop_chain_management.service;
 
+import com.example.coffee_shop_chain_management.dto.CreateAccountDTO;
 import com.example.coffee_shop_chain_management.entity.Account;
+import com.example.coffee_shop_chain_management.mapper.AccountMapper;
 import com.example.coffee_shop_chain_management.repository.AccountRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.AccessLevel;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+@RequiredArgsConstructor
+@FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
 @Service
 public class AccountService {
-    @Autowired
     private AccountRepository accountRepository;
+    private AccountMapper accountMapper;
+    private PasswordEncoder passwordEncoder;
 
-    public Account createAccount(Account account){
+    public Account createAccount(CreateAccountDTO accountDTO){
+        if(accountRepository.existsByUsername(accountDTO.getUsername())){
+            throw new RuntimeException("Username is already taken!");
+        }
+
+        Account account = accountMapper.toAccount(accountDTO);
+        account.setPassword(passwordEncoder.encode(accountDTO.getPassword()));
+
         return accountRepository.save(account);
     }
 
@@ -19,15 +34,15 @@ public class AccountService {
     }
 
     public Account getAccountByUsername(String username){
-        return accountRepository.findByUsername(username).get();
+        return accountRepository.findByUsername(username).orElse(null);
     }
 
     public Account getAccountByEmail(String email){
-        return accountRepository.findByEmail(email).get();
+        return accountRepository.findByEmail(email).orElse(null);
     }
 
-    public Account updateAccount(Account account){
-        return accountRepository.save(account);
+    public void updateAccount(Account account){
+        accountRepository.save(account);
     }
 
     public void deleteAccount(Account account){
@@ -35,6 +50,12 @@ public class AccountService {
     }
 
     public void deleteAccountById(Long id){
+        Account account = accountRepository.findById(id).orElse(null);
+
+        if(account == null){
+            throw new RuntimeException("Account not found!");
+        }
+
         accountRepository.deleteById(id);
     }
 }
